debug: 채팅내역 실시간 반영하기
