+++ src/types/chat/message.ts
@@ -1,0 +1,9 @@
+interface IMessage {
+  messageId: string;
+  roomId: string;
+  senderId: string;
+  senderNickname: string;
+  content: string;
+  timestamp: string;
+}
+export default IMessage;
